ML/requirements.txt
fastapi
uvicorn
numpy
tensorflow
scikit-learn
google-cloud-storage
Pillow
python-multipart